Add kics description (#314)
* add description field
